Implemented gifs, general restructuring
